% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_path.R
\name{clean_path}
\alias{clean_path}
\alias{backslash}
\alias{forwardslash}
\alias{bs}
\alias{fs}
\title{standardize paths and easy clipboard path updating}
\usage{
clean_path(path, backslash = T, force_endslash = F)

backslash(path)

forwardslash(path)

bs()

fs()
}
\arguments{
\item{path}{supply a path}

\item{backslash}{T/F backslashes or forwardslashes?}

\item{force_endslash}{T/F cap path with slash?}
}
\value{
updated path
}
\description{
paths often have extra slashes, or in opposing directions. These functions standardize slashes
}
\details{
standardizes slashes - forward or back

 This program is free software but it is provided WITHOUT WARRANTY
 and with ABSOLUTELY NO GUARANTEE of fitness or functionality for any purpose;
 you can redistribute it and/or modify it under the terms of the GNU
 General Public License as published by the Free Software Foundation;
 either version 2 of the License, or (at your option) any later version.

\cr

Revision History
\tabular{ll}{
1.0 \tab 2015 Jan 29 Roxygen header prepared \cr
}
}
\examples{

       clean_path("c:\\\\temp",F,F)
       clean_path("c:\\\\temp",T,F)
       clean_path("c:\\\\temp",T,T)

       clean_path("c:/temp/",F,F)
       clean_path("c:\\\\temp\\\\\\\\",T,F)
       clean_path("c:\\\\temp\\\\\\\\",T,T)




}
\seealso{
\code{\link{grep}}\cr \code{\link{gsub}}\cr
}
\author{
Jacob Strunk <jstrunk@fs.fed.us>
}
