% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.flextable.R
\name{color.flextable}
\alias{color.flextable}
\title{create colored flextable from input dataframe and write to ms word}
\usage{
color.flextable(
  x,
  colFields = NA,
  colRows = NA,
  cols = colorRamps::blue2red(40)[10:40],
  breaks = NA,
  breaksType = c("quantile", "equal"),
  outDocx = NA,
  preview = T,
  digits = 0,
  return = T
)
}
\arguments{
\item{x}{input data.frame}

\item{colFields}{numeric fields to color}

\item{colRows}{choose a subset of rows (numbers or row ids) to color}

\item{cols}{sequence of colors as from colorRamps::blue2red(40)}

\item{breaks}{(optional) sequence of break values to use in slicing x into colors}

\item{breaksType}{(optional) if breaks = NA, then use either quantile or equal interval breaks}

\item{outDocx}{(optional) provide location to write flex table as word document}

\item{preview}{should a temporary word document with flex table be opened}

\item{digits}{how many digits should be retained (round) for numeric values}

\item{return}{T/F should a flextable be returned?}
}
\value{
NULL or a flex table if return == T
}
\description{
create colored flextable from input dataframe and write to ms word
}
\details{
This program is free software but it is provided WITHOUT WARRANTY
 and with ABSOLUTELY NO GUARANTEE of fitness or functionality for any purpose;
 you can redistribute it and/or modify it under the terms of the GNU
 General Public License as published by the Free Software Foundation;
 either version 2 of the License, or (at your option) any later version.

\cr
Revision History
\tabular{ll}{
1.0 \tab 7/7/2020 function created \cr
1.1 \tab date and revisions.. \cr
}
}
\examples{
 <Delete and Replace>

}
\seealso{
\code{\link{print.flextable}}\cr \code{\link{flextable}}\cr \code{\link{quantile}}\cr \code{\link{save_as_docx}}\cr
}
\author{
Jacob Strunk <Jacob.strunk@usda.gov>
}
