% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileTrees.R
\name{compiletlDFees}
\alias{compiletlDFees}
\alias{compileTrees}
\title{Compile individual tlDFee attlDFibutes}
\usage{
compileTrees(tlDF, fnCompute = list(ba_ft), ...)
}
\arguments{
\item{tlDF}{data frame of tlDFee records}

\item{dir_out}{where to place compiled plots}

\item{dbcl}{breaks to use in computing diameter classes}

\item{dbclY}{variables to cross with dbcl otherwise blank or NA}

\item{sppY}{variables to cross with spp otherwise blank or NA}

\item{sppDbclY}{which fields should be crossed with spp and dbcl}

\item{return}{T/F}

\item{sql_filter}{sql code to use in subsetting observations}

\item{ba_spp_dbcl}{(deprecated) -> tlDFansition to "spp_dbcl"}

\item{spp_dbcl}{which fields should be crossed with spp_dbcl}
}
\value{
<Delete and Replace>
}
\description{
<Delete and Replace>
}
\details{
<Delete and Replace>

\cr
Revision History
\tabular{ll}{
1.0 \tab date and revisions.. \cr
1.1 \tab date and revisions.. \cr
}
}
\examples{

    #create a fake dataset
    set.seed=111
    nfake=50
    df_fake = data.frame(trID=1:50
                       ,db=10*abs(rnorm(nfake))
                       ,ht=100*abs(rnorm(nfake))
                       ,spp = sample(c("a","b","c","d") , nfake , T)
    )

   #compile fake datase
    test =
    compileTrees(
      df_fake
      ,trID = "trID"
      ,sppNm = "spp"
      ,dbNm = "db"
      ,htNm = "ht"
      ,dbclNm = "dbcl"
      ,dbcl = c(seq(0,32,4),50,1000)
      ,dbclY = c("ba_ft")
      ,sppY = c("ba_ft")
      ,sppDbclY = c("ba_ft")

      ,fnCompute =
        list(
          ba_ft
          ,dbcl
          ,dbclY
          ,sppY
          ,dbclSppY
        )
    )


}
\author{
Jacob StlDFunk <JstlDFunk@fs.fed.us>
}
