% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileTrees.R
\name{ba_ft}
\alias{ba_ft}
\alias{tpa}
\alias{tph}
\alias{dbcl}
\alias{dbclY}
\alias{sppY}
\alias{dbclSppY}
\title{Optional compilation function to be supplied in fnCompute list argument: fnCompute = list(ba_ft,...)}
\usage{
ba_ft(x, dbNm, ...)

tpa(x, acresNm = NA, nTreesNm = NA, ...)

tph(x, haNm = NA, nTreesNm = NA, ...)

dbcl(x, dbNm = "dbh", dbcl = c(seq(0, 32, 4), 50, 1000), dbclNm = "dbcl", ...)

dbclY(x, trID, dbclNm, dbclY, ...)

sppY(x, trID, sppY, sppNm, ...)

dbclSppY(x, trID, sppY, dbclNm, sppNm, ...)
}
\description{
Optional compilation function to be supplied in fnCompute list argument: fnCompute = list(ba_ft,...)

Optional compilation function to be supplied in fnCompute list argument: fnCompute = list(ba_ft,...)

Optional compilation function to be supplied in fnCompute list argument: fnCompute = list(ba_ft,...)

Optional compilation function to be supplied in fnCompute list argument: fnCompute = list(ba_ft,...)

Optional compilation function to be supplied in fnCompute list argument: fnCompute = list(ba_ft,...)

Optional compilation function to be supplied in fnCompute list argument: fnCompute = list(ba_ft,...)

Optional compilation function to be supplied in fnCompute list argument: fnCompute = list(ba_ft,...)
}
