% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_raster.R
\name{df_to_raster}
\alias{df_to_raster}
\title{Convert columns of metrics in a data.frame to rasters, requires}
\usage{
df_to_raster(
  df,
  colsxy = c("center_x", "center_y"),
  cols2Raster = c(colsSomeX),
  format = ".img",
  dirOut = "E:\\\\projects\\\\2017_NAIP\\\\rasters\\\\",
  crs = NA,
  nProc = 10,
  doDebug = F
)
}
\arguments{
\item{df}{a dataframe with xy and columns to use in making a rasters}

\item{colsxy}{names of xy columns}

\item{cols2Raster}{which columns to grab}

\item{format}{raster formate  e.g. .tif , .img etc}

\item{dirOut}{where to export rasters}

\item{crs}{proj4 string or other crs notation used by raster package (see writeRaster)}

\item{nProc}{number of cores to use in converting to raster}

\item{doDebug}{run in debug mode and only read 50k rows?}
}
\value{
<Delete and Replace>
}
\description{
<Delete and Replace>
}
\details{
<Delete and Replace>

\cr
Revision History
\tabular{ll}{
1.0 \tab 3/16/2020 Function created \cr
}
}
\examples{
 <Delete and Replace>

}
\seealso{
\code{\link{rasterFromXYZ}}\cr \code{\link{writeRaster}}\cr
}
\author{
Jacob Strunk <Jstrunk@fs.fed.us>
}
