% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_duplicates.R
\name{duplicated2}
\alias{duplicated2}
\title{This program is free software but it is provided WITHOUT WARRANTY
 and with ABSOLUTELY NO GUARANTEE of fitness or functionality for any purpose;
 you can redistribute it and/or modify it under the terms of the GNU
 General Public License as published by the Free Software Foundation;
 either version 2 of the License, or (at your option) any later version.}
\usage{
duplicated2(x, type = c("all", "first", "last"), value = F, invert = F)
}
\description{
This program is free software but it is provided WITHOUT WARRANTY
 and with ABSOLUTELY NO GUARANTEE of fitness or functionality for any purpose;
 you can redistribute it and/or modify it under the terms of the GNU
 General Public License as published by the Free Software Foundation;
 either version 2 of the License, or (at your option) any later version.
}
