% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_pdf.R
\name{fit_pdf}
\alias{fit_pdf}
\alias{pdf_compare}
\title{fit and compare pdfs}
\usage{
fit_pdf(
  y,
  wts = NA,
  bins = NA,
  breaks = NA,
  lims = range(c(y, bins)),
  MA = TRUE,
  MA_n = 3,
  plot = F
)

pdf_compare(
  fit_pdf1,
  fit_pdf2,
  plot = TRUE,
  legend = T,
  xlab = "bin",
  ylab = "y_dens",
  add = F,
  ...
)
}
\arguments{
\item{y}{vector of respone values}

\item{wts}{optional vector of weights for each y}

\item{bins}{bins to use in fitting pdf (pmf..)}

\item{lims}{upper and lower limits to compare across}

\item{MA}{T/F use moving average smoother?}

\item{MA_n}{number of bins to include in smoother}

\item{plot}{T/F plot two histograms?}

\item{fit_pdf1}{first pdf to use in comparison - object returned by fit_pdf()}

\item{fit_pdf2}{second pdf to use in comparison - object returned by fit_pdf()}

\item{add}{T/F histogram to existing plot? - in the case of multiple comparisons...}

\item{...}{additional arguments to plot}
}
\value{
fit_pdf: a pdf specially prepared to be handled by pdf_compare
}
\description{
fit and compare pdfs, three indices are reported which are similar to R-square, RMSE, and standard error
for the comparisons. One of the fits, the baseline, is treated as the population pdf (much like the population
mean when computing r-squared).

a pdf fitting function is specialized to enable comparison across a desired range
a function which computes fitting indices which enable comparison between fitted pdfs
}
\details{
fit multiple pdfs with fit_pdf, and specify the range of possibly y values across with comparisons should be made
then run pdf_compare to evaluate the relative performance of the fit

\cr

Revision History
\tabular{ll}{
1.0 \tab 4/1/2014 operation - header needs work \cr
}
}
\examples{

            pdf("c:/temp/test.pdf")
            x1=fit_pdf(rnorm(500,0,1),plot=F,bins=seq(-10,200,1))
            x2=fit_pdf(rnorm(500,0,1),plot=F,bins=seq(-10,200,1))
            pdf_compare(x1,x2,plot=T)
            x1=fit_pdf(rnorm(500,0,1),plot=F,bins=seq(100,200,1))
            x2=fit_pdf(rbeta(500,1,1),plot=F,bins=seq(100,200,1))
            pdf_compare(x1,x2,plot=T)
            dev.off()




}
\seealso{
\code{\link{another_function}}\cr \code{\link{yet_another_function}}\cr
}
\author{
Jacob Strunk <Jacob.strunk@dnr.wa.gov>
}
