% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFIA.R
\name{getFIA}
\alias{getFIA}
\title{Get FIA data from citrix (Make Sure You Are Using 32bit R)}
\usage{
getFIA(
  nims_channel,
  dbqrys = c(tr =
    "select * from fs_nims_pnwrs.nims_tree_vw where statecd > -1 and invyr > -1 and COUNTYCD < 10E6 and rownum < 50",
    pl_snp =
    "select * from fs_nims_pnwrs.nims_plotsnap_vw where statecd > -1 and invyr > -1 and COUNTYCD < 10E6 and rownum < 50",
    cond =
    "select * from fs_nims_pnwrs.nims_cond_vw where statecd > -1 and invyr > -1 and COUNTYCD < 10E6 and rownum < 50",
    coords =
    "select * from FS_NIMS_PNWRS.NIMS_GPS_PNWRS where statecd > -1 and SAMPLE_YEAR > -1 and COUNTYCD < 10E6 and rownum < 50",
    wts =
    "select * from fs_nims_pnwrs.nims_pop_expall_vw where statecd > -1 and invyr > -1 and COUNTYCD < 10E6 and rownum < 50"),
  dir_out = c(file.path("c:/data/RFIA/NIMS", format(Sys.Date())), NA),
  return = F,
  put_data_parent_env = F,
  stringsAsFactors = F
)
}
\arguments{
\item{nims_channel}{<Delete and Replace>}

\item{dbqrys}{oracle slq querys to grab tables. Must name queries (tr,pl_snp,cond,coords,wts)}

\item{dir_out}{where to save records for individual tables}

\item{return}{should the data be returned after the function call}

\item{put_data_parent_env}{should the data be placed in the parent environment where the call was made from}

\item{stringsAsFactors}{should strings be converted to factors}
}
\value{
nothing or a list of data frames
}
\description{
<Delete and Replace>
}
\details{
<Delete and Replace>


 This program is free software but it is provided WITHOUT WARRANTY
 and with ABSOLUTELY NO GUARANTEE of fitness or functionality for any purpose;
 you can redistribute it and/or modify it under the terms of the GNU
 General Public License as published by the Free Software Foundation;
 either version 2 of the License, or (at your option) any later version.



\cr
Revision History
\tabular{ll}{
1.0 \tab date and revisions.. \cr
}
}
\examples{

 #Make Sure You Are Using 32bit R

library(RODBC)
chn1 = odbcConnect("FIA01P")

getFIA(chn1)

res = getFIA(return = T)
head(res[["tr"]])

getFIA(return = T)
head(tr)

 #get an idea of what each table looks like
 getFIA(chn1
			,dbqrys=c(
				tr = 'fs_nims_pnwrs.nims_tree_vw where rownum < 50'
				,pl_snp = 'fs_nims_pnwrs.nims_plotsnap_vw where rownum < 50'
				,cond = 'fs_nims_pnwrs.nims_cond_vw where where rownum < 50'
				,coords = 'FS_NIMS_PNWRS.NIMS_GPS_PNWRS where rownum < 50'
				,wts = 'fs_nims_pnwrs.nims_pop_expall_vw where where rownum < 50'
			))


}
\seealso{
\code{\link{another_function}}\cr \code{\link{yet_another_function}}\cr
}
\author{
Jacob Strunk <Jstrunk@fs.fed.us>
}
