% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsubVec.R
\name{gsubVec}
\alias{gsubVec}
\title{gsub using a vector or pattern and replacement values}
\usage{
gsubVec(pattern, replacement, x, ...)
}
\arguments{
\item{pattern}{vector of patterns}

\item{replacement}{vector of replacements}

\item{...}{all other values as described in gsub}
}
\value{
<Delete and Replace>
}
\description{
<Delete and Replace>
}
\details{
<Delete and Replace>

\cr
Revision History
\tabular{ll}{
1.0 \tab 5/1/2020 created \cr
}
}
\examples{
 <Delete and Replace>

}
\seealso{
\code{\link{gsub}}\cr
}
\author{
Jacob Strunk <Jacob.strunk@usda.gov>
}
