% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs2.R
\name{pairs2}
\alias{pairs2}
\title{xy plot of variables in two dataframes}
\usage{
pairs2(x, y, str_max = 12, oneLine = F, mai_fig = c(0.3, 0.3, 0.3,
  0.1), mai_edge = c(0, 0, 0, 0), title = NA, lattice = F,
  pdf_out = NA, pdf_ht = 8, pdf_wd = 11, ...)
}
\arguments{
\item{x}{first data.frame}

\item{y}{second data.frame}

\item{str_max}{number of characters per name to display}

\item{oneLine}{T/F add one to one line?}

\item{mai_edge}{"mai" parameters for composite figure}

\item{...}{additional arguments to plot}

\item{tile}{optional figure title}
}
\value{
<Delete and Replace>
}
\description{
accepts two dataframes and prepares pairplots, results in fewer combinations than plot(xy_df)
}
\details{
<Delete and Replace>

\cr

Revision History
\tabular{ll}{
1.0 \tab 2015 Jan 29 Roxygen header prepared \cr
}
}
\examples{

pairs2( x = x_ht_debug ,  y = y1 , title = "Hello" )


}
\seealso{
\code{\link{plot}}\cr
}
\author{
Jacob Strunk <Jacob.strunk@dnr.wa.gov>
}
