% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableColorRamp.R
\name{tableColorRamp}
\alias{tableColorRamp}
\title{Makes a table with cells colored by relative rank (e.g. blue to red)}
\usage{
tableColorRamp(
  data,
  colorRamp = colorRamps::blue2red,
  nColors = 20,
  pdfOut = c("c:\\\\temp\\\\test.pdf", NA)[1],
  pdfHeight = 8,
  pdfWidth = 11,
  plotMar = c(8, 8, 8, 10),
  plotMain = "Table Of Values",
  legendPrecision = 0,
  ...
)
}
\arguments{
\item{data}{data frame of input values to plto}

\item{colorRamp}{color ramp function, e.g. colorRamps package or grDevices::colorRampPalette(c("blue", "red"))}

\item{nColors}{integer number of colors}

\item{pdfOut}{optional output pdf file, NA will result in plotting the R interface}

\item{pdfHeight}{width of output pdf}

\item{pdfWidth}{height of output pdf}

\item{plotMar}{plot margins}

\item{plotMain}{main argument to image function}

\item{legendPrecision}{decimal precision to use for legend object}

\item{...}{additional arguments to image function}
}
\value{
NULL - makes a plot or writes plot to pdf file
}
\description{
Makes a table with cells colored by relative rank (e.g. blue to red)
}
\details{
Uses the image function to make a table of colors and writes input data values to color cells

 This program is free software but it is provided WITHOUT WARRANTY
 and with ABSOLUTELY NO GUARANTEE of fitness or functionality for any purpose;
 you can redistribute it and/or modify it under the terms of the GNU
 General Public License as published by the Free Software Foundation;
 either version 2 of the License, or (at your option) any later version.

\cr
Revision History
\tabular{ll}{
1.0 \tab 6/28/2020 Created and added to package\cr

}
}
\examples{

   test = data.frame(matrix(rnorm(100)*100,nrow = 10))
   tableColorRamp(test)


}
\seealso{
\code{\link{image}}\cr
}
\author{
Jacob Strunk <Jacob.strunk@usda.gov>
}
