% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileTrees.R
\name{compileTrees}
\alias{compileTrees}
\title{Compile individual tree attributes}
\usage{
compileTrees(
  tr,
  fnArgs = list(trID = "CN", sppNm = "SPGRPCD", dbNm = "DIA", dbclNm = "dbcl", dbcl =
    c(seq(0, 32, 4), 50, 1000), dbclY = c("ba_ft"), sppY = c("ba_ft"), sppDbclY =
    c("ba_ft")),
  fnCompute = defComputeTL
)
}
\arguments{
\item{tr}{data frame of tree records}

\item{dir_out}{where to place compiled plots}

\item{dbcl}{breaks to use in computing diameter classes}

\item{dbclY}{variables to cross with dbcl otherwise blank or NA}

\item{sppY}{variables to cross with spp otherwise blank or NA}

\item{sppDbclY}{which fields should be crossed with spp and dbcl}

\item{return}{T/F}

\item{sql_filter}{sql code to use in subsetting observations}

\item{ba_spp_dbcl}{(deprecated) -> transition to "spp_dbcl"}

\item{spp_dbcl}{which fields should be crossed with spp_dbcl}
}
\value{
<Delete and Replace>
}
\description{
<Delete and Replace>
}
\details{
<Delete and Replace>

\cr
Revision History
\tabular{ll}{
1.0 \tab date and revisions.. \cr
1.1 \tab date and revisions.. \cr
}
}
\examples{
 <Delete and Replace>


}
\author{
Jacob Strunk <Jstrunk@fs.fed.us>
}
