% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileTrees.R
\name{compileTrees}
\alias{compileTrees}
\alias{ba_ft}
\alias{dbcl}
\alias{dbclY}
\alias{sppY}
\alias{dbclSppY}
\title{Facilitate computation of tree level attributes (ba, volume, etc.)}
\usage{
compileTrees(
  tlDF,
  fnCompute = list(tpa, ba_ft, dbcl, dbclY, sppY, dbclSppY),
  ...
)

ba_ft(x, dbNm, ...)

dbcl(x, dbNm = "dbh", dbcl = c(seq(0, 32, 4), 50, 1000), dbclNm = "dbcl", ...)

dbclY(x, trID, dbclNm, dbclY, ...)

sppY(x, trID, sppY, sppNm, ...)

dbclSppY(x, trID, sppY, dbclNm, sppNm, ...)
}
\arguments{
\item{tlDF}{data frame of tree records}

\item{fnCompute}{sequential list of functions to
apply to tree data, earlier results (e.g. ba) are available to later
functions. Every function should accept an elipsis}

\item{...}{arguments to functions in fnCompute}
}
\value{
<Delete and Replace>
}
\description{
<Delete and Replace>
}
\details{
<Delete and Replace>

 This program is free software but it is provided WITHOUT WARRANTY
 and with ABSOLUTELY NO GUARANTEE of fitness or functionality for any purpose;
 you can redistribute it and/or modify it under the terms of the GNU
 General Public License as published by the Free Software Foundation;
 either version 2 of the License, or (at your option) any later version.

\cr
Revision History
\tabular{ll}{
1.0 \tab 6/10/2020 Created  \cr
}
}
\examples{

    #create a fake dataset
    set.seed=111
    nfake=50
    df_fake = data.frame(trID=1:50
                       ,db=10*abs(rnorm(nfake))
                       ,ht=100*abs(rnorm(nfake))
                       ,spp = sample(c("a","b","c","d") , nfake , T)
    )

   #compile fake datase
    test =
    compileTrees(
      df_fake
      ,trID = "trID"
      ,sppNm = "spp"
      ,dbNm = "db"
      ,htNm = "ht"
      ,dbclNm = "dbcl"
      ,dbcl = c(seq(0,32,4),50,1000)
      ,dbclY = c("ba_ft")
      ,sppY = c("ba_ft")
      ,sppDbclY = c("ba_ft")

      ,fnCompute =
        list(
          ba_ft
          ,dbcl
          ,dbclY
          ,sppY
          ,dbclSppY
        )
    )


}
\author{
Jacob Strunk <Jstrunk@fs.fed.us>
}
